chore: changed createDisposables to return an instance of a class for better docs
