+++ packages/patterns/src/disposables/index.ts
@@ -1,4 +1,4 @@
-export * from './disposal-groups';
+export * from './create-disposables';
 export * from './disposable';
 export type { DisposableItem } from './disposables-group';
 export { DisposalGroup, GroupConstraints } from './constraints';
